+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- General
+  - Fix move-to-let/expand-let bug for multi-arity functions #1283
 
 - General
   - Bump clj-kondo to `2023.09.07`.